chore: fix circular dep for config
